# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lohikeitto",
    "description": "Service catalog API with search, logos, and categories. Aggregates data from a local database and external brand APIs (Brandfetch, logo.dev).",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://soup.uha.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health check",
        "description": "Returns the service status and database connectivity. Used by orchestrators (Docker, k8s) for liveness probes.",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service is degraded (database unreachable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search services",
        "description": "Search for services across inhouse database and external APIs (Brandfetch, logo.dev, App Store). Results are deduplicated by domain with priority: inhouse > appstore > brandfetch > logo.dev. Inhouse (curated) services may have multiple domains; external results always have one.",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search string (required, non-empty)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "description": "Comma-separated sources: `inhouse`, `brandfetch`, `logodev`, `appstore`. Aliases: `external` (brandfetch + logodev + appstore), `mobile` (appstore), `all` (default)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  }
                },
                "example": [
                  {
                    "domains": [
                      "adguard.com"
                    ],
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "logo_url": "https://s3.uha.app/logos/adguard.webp",
                    "name": "AdGuard",
                    "source": "inhouse"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Invalid input (empty or missing `q`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/service/{lookup}": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get a service by ID or domain",
        "description": "Look up a service by UUID or domain name. If not found locally, searches external sources (Brandfetch, logo.dev, App Store) and adds the result to the approval queue.",
        "operationId": "get",
        "parameters": [
          {
            "name": "lookup",
            "in": "path",
            "description": "Service UUID or domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "status",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "example": 400,
            "minimum": 0
          },
          "message": {
            "type": "string",
            "description": "Error description",
            "example": "q is required"
          },
          "status": {
            "type": "string",
            "description": "Always `error`",
            "example": "error"
          }
        },
        "example": {
          "code": 400,
          "message": "q is required",
          "status": "error"
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "db"
        ],
        "properties": {
          "db": {
            "type": "boolean",
            "description": "Whether the database is reachable"
          },
          "message": {
            "type": "string",
            "description": "Human-readable status message",
            "example": "Lohikeitto is ready"
          },
          "status": {
            "type": "string",
            "description": "Service status: `ok` or `degraded`",
            "example": "ok"
          }
        },
        "example": {
          "db": true,
          "message": "Lohikeitto is ready",
          "status": "ok"
        }
      },
      "SearchResult": {
        "type": "object",
        "required": [
          "id",
          "logo_url",
          "name",
          "domains",
          "source"
        ],
        "properties": {
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Service domains. Inhouse (curated) results may have multiple; external results always have one."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "For inhouse results — UUID v4 from DB; for external — deterministic UUID v5 from first domain"
          },
          "logo_url": {
            "type": "string",
            "description": "Logo image URL",
            "example": "https://s3.uha.app/logos/adguard.webp"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "AdGuard"
          },
          "source": {
            "type": "string",
            "description": "Result source: `inhouse`, `brandfetch`, `logo.dev`, or `appstore`",
            "example": "inhouse"
          }
        },
        "example": {
          "domains": [
            "adguard.com"
          ],
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "logo_url": "https://s3.uha.app/logos/adguard.webp",
          "name": "AdGuard",
          "source": "inhouse"
        }
      },
      "ServiceResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "domains",
          "alternative_names",
          "tags",
          "verified",
          "colors",
          "social_links",
          "logo_url"
        ],
        "properties": {
          "alternative_names": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bundle_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "colors": {},
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "logo_url": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ref_link": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": "string"
          },
          "social_links": {},
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "verified": {
            "type": "boolean"
          }
        },
        "example": {
          "category": "security",
          "colors": {
            "primary": "#000000"
          },
          "domains": [
            "adguard.com"
          ],
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "logo_url": "https://s3.uha.app/logos/adguard.webp",
          "name": "AdGuard",
          "ref_link": null,
          "slug": "adguard",
          "verified": true
        }
      }
    }
  },
  "tags": [
    {
      "name": "Search",
      "description": "Service search across local and external sources"
    },
    {
      "name": "Services",
      "description": "Service details"
    },
    {
      "name": "System",
      "description": "Health checks and diagnostics"
    }
  ]
}